refactor(bayn): separate market strategy planning pipeline
