+++ services/bayn/src/market-data/errors.ts
@@ -1,6 +1,7 @@
 import { isSqlError } from 'effect/unstable/sql/SqlError'
 
 import { OperationalError, operationalError, retryableOperationalError } from '../errors'
+import { isMarketDataVerificationError, renderMarketDataVerificationError } from '../market-data-verification'
 
 export const marketDataOperationError = (
   operation: 'check' | 'inspect' | 'inspect-publication' | 'load',
@@ -8,6 +9,18 @@ export const marketDataOperationError = (
   cause: unknown,
 ): OperationalError => {
   if (cause instanceof OperationalError) return cause
-  const makeError = isSqlError(cause) && !cause.isRetryable ? operationalError : retryableOperationalError
-  return makeError('market-data', operation, message, cause)
+  if (isSqlError(cause)) {
+    const makeError = cause.isRetryable ? retryableOperationalError : operationalError
+    return makeError('market-data', operation, message, cause)
+  }
+  if (isMarketDataVerificationError(cause)) {
+    return new OperationalError({
+      component: 'market-data',
+      operation,
+      message: renderMarketDataVerificationError(cause),
+      retryable: false,
+      cause,
+    })
+  }
+  return retryableOperationalError('market-data', operation, message, cause)
 }
